Refactor analysis into new namespace
