+++ bundles/org.dataflowanalysis.analysis/META-INF/MANIFEST.MF
@@ -9,16 +9,11 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="3.24.0",
  org.eclipse.core.resources;bundle-version="3.16.0";visibility:=reexport,
  tools.mdsd.library.standalone.initialization;bundle-version="0.2.0";visibility:=reexport,
  tools.mdsd.library.standalone.initialization.log4j;bundle-version="0.2.0";visibility:=reexport,
- tools.mdsd.library.standalone.initialization.emfprofiles;bundle-version="0.2.0";visibility:=reexport,
  org.palladiosimulator.pcm;bundle-version="5.1.0";visibility:=reexport,
- org.modelversioning.emfprofile;bundle-version="5.1.0";visibility:=reexport,
- org.modelversioning.emfprofile.registry;bundle-version="5.1.0";visibility:=reexport,
- org.palladiosimulator.dataflow.dictionary.characterized;bundle-version="5.1.0";visibility:=reexport,
- org.palladiosimulator.dataflow.confidentiality.pcm.dddsl;bundle-version="5.1.0";visibility:=reexport,
  org.apache.log4j;bundle-version="1.2.15";visibility:=reexport,
- org.palladiosimulator.dataflow.confidentiality.pcm.model.profile;bundle-version="5.1.0";visibility:=reexport,
- org.palladiosimulator.mdsdprofiles.api;bundle-version="4.3.0";visibility:=reexport,
- org.palladiosimulator.dataflow.nodecharacteristics;bundle-version="0.1.0"
+ org.dataflowanalysis.pcm.extension.dictionary.characterized;bundle-version="5.1.0";visibility:=reexport,
+ org.dataflowanalysis.pcm.extension.dddsl;bundle-version="5.1.0";visibility:=reexport,
+ org.dataflowanalysis.pcm.extension.nodecharacteristics;bundle-version="0.1.0"
 Export-Package: org.dataflowanalysis.analysis.builder,
  org.dataflowanalysis.analysis.builder.pcm,
  org.dataflowanalysis.analysis.characteristics,
