+++ bundles/org.dataflowanalysis.analysis/src/org/dataflowanalysis/analysis/characteristics/CharacteristicValue.java
@@ -1,7 +1,7 @@
 package org.dataflowanalysis.analysis.characteristics;
 
-import org.palladiosimulator.dataflow.dictionary.characterized.DataDictionaryCharacterized.EnumCharacteristicType;
-import org.palladiosimulator.dataflow.dictionary.characterized.DataDictionaryCharacterized.Literal;
+import org.dataflowanalysis.pcm.extension.dictionary.characterized.DataDictionaryCharacterized.EnumCharacteristicType;
+import org.dataflowanalysis.pcm.extension.dictionary.characterized.DataDictionaryCharacterized.Literal;
 
 public record CharacteristicValue(EnumCharacteristicType characteristicType, Literal characteristicLiteral) {
