+++ bundles/org.dataflowanalysis.analysis/src/org/dataflowanalysis/analysis/characteristics/node/PCMNodeCharacteristicsCalculator.java
@@ -10,16 +10,16 @@ import org.apache.log4j.Logger;
 import org.dataflowanalysis.analysis.characteristics.CharacteristicValue;
 import org.dataflowanalysis.analysis.resource.ResourceProvider;
 import org.dataflowanalysis.analysis.utils.pcm.PCMQueryUtils;
-import org.palladiosimulator.dataflow.confidentiality.pcm.model.confidentiality.characteristics.EnumCharacteristic;
-import org.palladiosimulator.dataflow.confidentiality.pcm.model.confidentiality.repository.OperationalDataStoreComponent;
-import org.palladiosimulator.dataflow.dictionary.characterized.DataDictionaryCharacterized.Literal;
-import org.palladiosimulator.dataflow.nodecharacteristics.nodecharacteristics.AbstractAssignee;
-import org.palladiosimulator.dataflow.nodecharacteristics.nodecharacteristics.AssemblyAssignee;
-import org.palladiosimulator.dataflow.nodecharacteristics.nodecharacteristics.Assignments;
-import org.palladiosimulator.dataflow.nodecharacteristics.nodecharacteristics.NodeCharacteristicsFactory;
-import org.palladiosimulator.dataflow.nodecharacteristics.nodecharacteristics.NodeCharacteristicsPackage;
-import org.palladiosimulator.dataflow.nodecharacteristics.nodecharacteristics.RessourceAssignee;
-import org.palladiosimulator.dataflow.nodecharacteristics.nodecharacteristics.UsageAsignee;
+import org.dataflowanalysis.pcm.extension.model.confidentiality.characteristics.EnumCharacteristic;
+import org.dataflowanalysis.pcm.extension.model.confidentiality.repository.OperationalDataStoreComponent;
+import org.dataflowanalysis.pcm.extension.dictionary.characterized.DataDictionaryCharacterized.Literal;
+import org.dataflowanalysis.pcm.extension.nodecharacteristics.nodecharacteristics.AbstractAssignee;
+import org.dataflowanalysis.pcm.extension.nodecharacteristics.nodecharacteristics.AssemblyAssignee;
+import org.dataflowanalysis.pcm.extension.nodecharacteristics.nodecharacteristics.Assignments;
+import org.dataflowanalysis.pcm.extension.nodecharacteristics.nodecharacteristics.NodeCharacteristicsFactory;
+import org.dataflowanalysis.pcm.extension.nodecharacteristics.nodecharacteristics.NodeCharacteristicsPackage;
+import org.dataflowanalysis.pcm.extension.nodecharacteristics.nodecharacteristics.RessourceAssignee;
+import org.dataflowanalysis.pcm.extension.nodecharacteristics.nodecharacteristics.UsageAsignee;
 import org.palladiosimulator.pcm.allocation.Allocation;
 import org.palladiosimulator.pcm.allocation.AllocationPackage;
 import org.palladiosimulator.pcm.core.composition.AssemblyContext;
