+++ releng/org.dataflowanalysis.analysis.targetplatform/org.dataflowanalysis.analysis.targetplatform.targetplatform.target
@@ -53,12 +53,11 @@
 			<repository location="https://updatesite.palladio-simulator.com/palladio-thirdparty-emfprofiles/releases/latest/"/>
 		</location>
 		<location includeAllPlatforms="false" includeConfigurePhase="true" includeMode="planner" includeSource="false" type="InstallableUnit" refresh="true">
-			<unit id="org.palladiosimulator.dataflow.dictionary.characterized" version="0.0.0"/>
-			<unit id="org.palladiosimulator.dataflow.confidentiality.pcm.dddsl" version="0.0.0"/>
-			<unit id="org.palladiosimulator.dataflow.confidentiality.pcm.model" version="0.0.0"/>
-			<unit id="org.palladiosimulator.dataflow.confidentiality.pcm.model.profile" version="0.0.0"/>
-			<unit id="org.palladiosimulator.dataflow.nodecharacteristics" version="0.0.0"/>
-			<repository location="https://updatesite.palladio-simulator.com/palladio-addons-dataflowconfidentiality-metamodel/nightly/"/>
+			<unit id="org.dataflowanalysis.pcm.extension.dictionary.characterized" version="0.0.0"/>
+			<unit id="org.dataflowanalysis.pcm.extension.dddsl" version="0.0.0"/>
+			<unit id="org.dataflowanalysis.pcm.extension.model" version="0.0.0"/>
+			<unit id="org.dataflowanalysis.pcm.extension.nodecharacteristics" version="0.0.0"/>
+			<repository location="https://dataflowanalysis.github.io/updatesite/nightly/pcm-dataflowanalysis-extension"/>
 		</location>
 		<location includeAllPlatforms="false" includeConfigurePhase="true" includeMode="planner" includeSource="false" type="InstallableUnit">
 			<unit id="org.palladiosimulator.indirections" version="0.0.0"/>
